Refactor DOCX table extraction into reusable module
